Function learn_complete_beginners_guide::memory_mng::borrowing::borrowing_fn

pub fn borrowing_fn()

Borrowing

Borrowing

Ony one variable can own a piece of memory

Variables can borrow ownership to other pieces of memory.

let a = 6;
let b = &b; //a reference

println!("{}", *b); //need to add star to use value of the father in this case the a  nad not the reference

a += 2 //error

**When the reference is destroy the ownership is return to the father in this case a. **

THe borrow has to match the mutability

Not use

let mut v = vec![1, 2, 3, 4, 5];
for i in &v {
println!("for loop : {}", i);
v.push(6); mutable borrow occurs here
}